requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
boto3==1.26.155
python-dotenv==1.0.0
requests==2.31.0
PyJWT==1.7.1
python-jose==3.3.0
python-multipart==0.0.5
bcrypt==3.2.2
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
tensorflow==2.13.0